refactor: form validation
