+++ frontend/errors/errorMessages.ts
@@ -2,11 +2,10 @@ interface ErrorsI {
   [idx: string]: string;
 }
 
-export const errors: ErrorsI = {
+const ErrorMessages: ErrorsI = {
+  DEFAULT: "An error occurred. Please check your internet connection.",
   EMAIL_EXISTS:
     "Registration failed. Please check if you are already registered, otherwise contact support for more info",
 };
 
-export const getMessageFromErrorCode = (errorCode: string): string => {
-  return errors[errorCode];
-};
+export default ErrorMessages;
