+++ frontend/types/compoundFieldSet.tsx
@@ -1,0 +1,5 @@
+export interface CompoundFieldSetType {
+  label: string;
+  id: string;
+  inputType: string;
+}
